+++ client/demo/siteapi.js
@@ -1,3 +1,5 @@
+import uuid from 'uuid/v1';
+
 import {Site} from '../model/site';
 
 import {BaseAPI} from './baseapi';
@@ -21,5 +23,31 @@ module.exports = class extends BaseAPI {
         });
       });
     });
+  }
+
+  add_user(email, name, admin, cb) {
+    this.transaction('users', 'readwrite', (trans) => {
+      const users_store = trans.objectStore('users');
+      this.r(users_store.add(
+        {id: uuid(), email: email, name: name, admin: admin, nick:''}), () => {
+          this.all(users_store.openCursor(), (users) => {
+            this.update(trans, {site: {users: users}}, cb);
+          }, cb);
+        }, cb);
+    }, cb);
+  }
+
+  change_user_type(uid, admin, cb) {
+    this.transaction('users', 'readwrite', (trans) => {
+      const users_store = trans.objectStore('users');
+      this.r(users_store.get(uid), (user) => {
+        user.admin = admin;
+        this.r(users_store.put(user), () => {
+          this.all(users_store.openCursor(), (users) => {
+            this.update(trans, {site: {users: users}}, cb);
+          }, cb);
+        }, cb);
+      }, cb);
+    }, cb);
   }
 };